requirements.txt
h5py==2.7.1
keras==2.1.5
tensorflow==1.7.0
pillow==5.1.0
numpy==1.14.2
requests==2.18.4